+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,4 +1,6 @@
-# 1.5.0
+# 1.5.0 (unreleased)
+- Chart component: remove d3-array dependency.
+- Chart component: fix display when there is no data.
 - Improves display of charts where all values are 0.
 - Fix X-axis labels in hourly bar charts.
 - New `<Search>` prop named `showClearButton`, that will display a 'Clear' button when the search box contains one or more tags.